Fix meterReadingDayCompleteness as float not int
2026-06-19 00:01:26,682 - INFO - meterReadingDayCompleteness: 0.941

+++ python_frank_energie/models.py
@@ -2141,7 +2141,7 @@ class MonthSummary:
         if not isinstance(expected_costs, (int, float, type(None))):
             raise RequestException("Invalid expectedCosts")
 
-        if not isinstance(completeness, int):
+        if not isinstance(completeness, float):
             raise RequestException("Invalid meterReadingDayCompleteness")
 
         if not isinstance(gas_excluded, bool):
